third_party/blink/web_tests/html/selectmenu/selectmenu-slot-warning-listbox.html

<!DOCTYPE html>
<html lang="en">
<script src='../../resources/testharness.js'></script>
<script src='../../resources/testharnessreport.js'></script>
<script src="../../resources/testdriver.js"></script>
<script src="../../resources/testdriver-actions.js"></script>
<script src="../../resources/testdriver-vendor.js"></script>

<!-- No console warnings: listbox was replaced. -->
<selectmenu>
  <div slot="listbox">
    <div popover behavior="listbox"></div>
  </div>
</selectmenu>

<!-- One console warning: listbox was not replaced. -->
<selectmenu>
  <div slot="listbox">
    <div popover></div>
  </div>
</selectmenu>

<script>
  setup({ single_test: true });
  // The purpose of this test to to ensure that the correct
  // console warnings are logged, via comparison with the expectations
  // file. So, nothing to test here.
  done();
</script>
