third_party/blink/web_tests/html/selectmenu/selectmenu-slot-warning-listbox-expected.txt

CONSOLE WARNING: A <selectmenu>'s default listbox was removed and a new one was not provided. This <selectmenu> will not be fully functional.
This is a testharness.js-based test.
PASS selectmenu-slot-warning-listbox
Harness: the test ran to completion.
